добавление автора в статьи, доработка методов контроллера

+++ src/main/java/com/example/nauchki/service/PostService.java
@@ -1,12 +1,14 @@
 package com.example.nauchki.service;
 
 import com.example.nauchki.exceptions.ResourceNotFoundException;
+import com.example.nauchki.jwt.TokenUtils;
 import com.example.nauchki.mapper.PostMapper;
 import com.example.nauchki.model.Post;
 import com.example.nauchki.model.dto.PostDto;
 import com.example.nauchki.repository.PostRepo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Example;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -23,6 +25,7 @@ public class PostService {
     private final PostRepo postRepo;
     private final FileService fileService;
     private final PostMapper postMapper;
+    private final TokenUtils tokenUtils;
 
     public List<PostDto> getPost(Post filter) {
         return postRepo.findAll(Example.of(filter)).stream().map(postMapper::toDto).collect(Collectors.toList());
@@ -36,12 +39,19 @@ public class PostService {
         return postRepo.findByTag(tag).stream().map(postMapper::toDto).collect(Collectors.toList());
     }
 
+    //статью может удалить либо автор, либо пользователь с ролью SUPERADMIN
     @Transactional
     public boolean deletePost(Long id, Principal principal) {
+
         Post post= postRepo.findById(id).orElseThrow(
                 ()->new ResourceNotFoundException("Статья с id '" + id + "' не найдена"));
-        postRepo.delete(post);
+        boolean permition = hasPermitionForDelete(post, principal.getName());
+        if(!permition){
+            throw new AccessDeniedException("Удалить статью может только автор, или администратор");
+        }
+
         delAllImages(id, principal);
+        postRepo.delete(post);
         return true;
     }
 
@@ -62,6 +72,10 @@ public class PostService {
         if (file != null && !file.getOriginalFilename().isEmpty()) {
             Optional<Post> post = postRepo.findById(postId);
             Post postModel = post.orElseThrow(()->new ResourceNotFoundException("Post '" + postId + "' not found"));
+            boolean permition = hasPermitionForDelete(postModel, principal.getName());
+            if(!permition){
+                throw new AccessDeniedException("Загрузить изображение может только автор, или администратор");
+            }
             String path = fileService.saveAttachedFile(file, postModel, tags, description);
             postRepo.save(postModel);
             return path;
@@ -74,6 +88,10 @@ public class PostService {
         if (file != null && !file.getOriginalFilename().isEmpty()) {
             Optional<Post> post = postRepo.findById(postId);
             Post postModel = post.orElseThrow(()->new ResourceNotFoundException("Post '" + postId + "' not found"));
+            boolean permition = hasPermitionForDelete(postModel, principal.getName());
+            if(!permition){
+                throw new AccessDeniedException("Загрузить изображение может только автор, или администратор");
+            }
             String path = fileService.saveAttachedFile(file, postModel);
             postRepo.save(postModel);
             return path;
@@ -85,6 +103,11 @@ public class PostService {
     public void delImage(Long postId, Long imgid, Principal principal) {
         Optional<Post> post = postRepo.findById(postId);
         Post postModel = post.orElseThrow(()->new ResourceNotFoundException("Post '" + postId + "' not found"));
+        boolean permition = hasPermitionForDelete(postModel, principal.getName());
+        if(!permition){
+            throw new AccessDeniedException("Удалить изображение может только автор, или администратор");
+        }
+
         boolean fileConsists = postModel.getFiles().stream()
                 .anyMatch(v-> v.getId().equals(imgid));
         if(!fileConsists){
@@ -98,9 +121,24 @@ public class PostService {
     public void delAllImages(Long postId, Principal principal) {
         Optional<Post> post = postRepo.findById(postId);
         Post postModel = post.orElseThrow(()->new ResourceNotFoundException("Post '" + postId + "' not found"));
+        boolean permition = hasPermitionForDelete(postModel, principal.getName());
+        if(!permition){
+            throw new AccessDeniedException("Удалить изображения может только автор, или администратор");
+        }
         if(fileService.deleteAllAttachedFiles(postModel)){
             postRepo.save(postModel);
         }
+    }
+
+    private boolean hasPermitionForDelete(Post post, String userName){
+        boolean permition = false;
+        List<String> roles = tokenUtils.getRoles();
+        if(roles.contains("SUPERADMIN")){
+            permition = true;
+        }else if(post.getAuthor().getEmail().equals(userName)){
+            permition = true;
+        }
+        return permition;
     }
 
     /*public boolean addPost(User user, String text, String tag, MultipartFile file){
